renamed the package and components to mainatin consistency
