--- instana-rust-exporter/examples/matrix-multiplication/Cargo.toml
@@ -1,28 +1,0 @@
-[package]
-name = "matrix_multiplication"
-version = "1.0.0"
-edition = "2021"
-
-[dependencies]
-actix-web = "4.11.0"
-serde = { version = "1.0", features = ["derive"] }
-features = "0.10.0"
-serde_json = "1.0.140"
-actix-cors = "0.7.1"
-opentelemetry = "0.29.1"
-opentelemetry_sdk = "0.29.0"
-opentelemetry-otlp={ version = "0.29.0", features = ["grpc-tonic", "gzip-tonic"] }
-tokio = "1.45.1"
-tracing = "0.1.41"
-tracing-subscriber = "0.3.19"
-tracing-opentelemetry = "0.31.0"
-#instana_opentelemetry_sdk = { git = "ssh://git@github.ibm.com/instana/instana-exporter-rust.git" }
-instana_opentelemetry_sdk = { path = "../.." }
-reqwest = { version = "0.11.24", features = ["json"] }
-tracing-actix-web = "0.7.18"
-actix-web-opentelemetry = "0.22.0"
-
-# Single binary target
-[[bin]]
-name = "matrix_multiplication"
-path = "src/main.rs"
