--- instana-rust-exporter/examples/matrix-printer/Cargo.toml
@@ -1,23 +1,0 @@
-[package]
-name = "matrix-printer"
-version = "1.0.0"
-edition = "2021"
-
-[dependencies]
-actix-cors = "0.7.1"
-actix-web = "4.11.0"
-actix-web-opentelemetry = "0.22.0"
-opentelemetry = "0.29.1"
-opentelemetry_sdk = "0.29.0"
-reqwest = "0.11.24"
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
-
-#instana_opentelemetry_sdk = { git = "ssh://git@github.ibm.com/instana/instana-exporter-rust.git" }
-instana_opentelemetry_sdk = { path = "../.." }
-tracing-actix-web = "0.7.18"
-
-# Single binary target
-[[bin]]
-name = "matrix-printer"
-path = "src/main.rs"
