--- instana-rust-exporter/src/exporter/mod.rs
@@ -1,5 +1,0 @@
-pub mod defs;
-pub mod exporter;
-pub mod instana_span;
-pub mod serialize_span;
-pub mod span_data;
