--- instana-rust-exporter/src/lib.rs
@@ -1,2 +1,0 @@
-pub mod exporter;
-pub mod propagator;
