--- instana-rust-exporter/src/propagator/mod.rs
@@ -1,5 +1,0 @@
-mod composite;
-mod propagator;
-
-pub use composite::CompositePropagator;
-pub use propagator::Propagator;
