+++ opentelemetry-instana/Cargo.toml
@@ -1,0 +1,26 @@
+[package]
+name = "opentelemetry_instana"
+version = "1.0.0"
+edition = "2021"
+license = "Apache-2.0"
+publish = true
+description="Instana Opentelemetry Rust SDK"
+
+
+[dependencies]
+bytes = "1.10.1"
+features = "0.10.0"
+http = "1.3.1"
+json = "0.12.4"
+opentelemetry = "0.29.1"
+opentelemetry-http = { version = "0.29.0", features = ["hyper","reqwest"] }
+opentelemetry-proto = "0.29.0"
+opentelemetry_sdk = "0.29.0"
+reqwest = { version = "0.12.15", features = ["blocking"] }
+serde = "1.0.219"
+serde_json = "1.0.140"
+thiserror = "2.0.12"
+tokio = { version = "1.45.0", features = ["macros", "rt-multi-thread"]} 
+url = "2.5.4"
+wiremock = "0.6.3"
+anyhow = "1.0"
