+++ opentelemetry-instana/examples/run_servers.sh
