+++ opentelemetry-instana/sonar-project.properties
