+++ opentelemetry-instana/src/exporter/span_data.rs
