+++ opentelemetry-instana/src/lib.rs
@@ -1,0 +1,5 @@
+pub mod exporter;
+pub mod propagator;
+
+pub use exporter::{InstanaExporter,InstanaExporterOptions};
+pub use propagator::{InstanaPropagator};
