+++ opentelemetry-instana/whitesource.config
